+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.1.1] 2025-05-14
+
 ### Added
 
 - Issue [#15](https://github.com/m-marini/qucomp/issues/15): Add function mapping state